requirements.txt
# Core dependencies
click>=8.0
rich>=13.0
colorama>=0.4

# HTTP requests
requests>=2.28
aiohttp>=3.8
aiohttp-socks>=0.8

# DNS and WHOIS
dnspython>=2.3
python-whois>=0.8

tldextract>=3.4

# Optional: For better JSON output
orjson>=3.8

# Development
pytest>=7.0
pytest-asyncio>=0.21
black>=23.0
flake8>=6.0
mypy>=1.0